WIP Extraction of nuget-upgrade
